feat(slo): improve burn rate panel (#197339)
(cherry picked from commit a5b266c45b7db8821281a2325d919b6fdf679618)
