--- x-pack/plugins/observability_solution/slo/public/components/slo/burn_rate/alert_details/components/custom_panels/custom_kql/helpers/log_rate_analysis_query.test.ts
@@ -1,195 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { getESQueryForLogRateAnalysis } from './log_rate_analysis_query';
-import { KQLCustomIndicator } from '@kbn/slo-schema';
-
-describe('buildEsQuery', () => {
-  const testData = [
-    {
-      title: 'rule with good query (as KQL), no optional filter and no total',
-      params: {
-        filter: '',
-        total: '',
-        good: 'http.response.status_code < 500',
-      },
-    },
-    {
-      title: 'rule with good query (as filter), no optional filter and no total',
-      params: {
-        filter: '',
-        total: '',
-        good: {
-          kqlQuery: '',
-          filters: [
-            {
-              $state: {
-                store: 'appState',
-              },
-              meta: {
-                field: 'http.response.status_code',
-                negate: false,
-                alias: null,
-                index: 'kbn-data-forge-fake_stack.admin-console-*-id',
-                disabled: false,
-                params: {
-                  lt: '500',
-                },
-                type: 'range',
-                key: 'http.response.status_code',
-              },
-              query: {
-                range: {
-                  'http.response.status_code': {
-                    lt: '500',
-                  },
-                },
-              },
-            },
-          ],
-        },
-      },
-    },
-    {
-      title: 'rule with good query (as KQL), with optional filter (as KQL) and not total filter',
-      params: {
-        total: '',
-        good: 'http.response.status_code < 500',
-        filter: 'host.name: admin-console.prod.001',
-      },
-    },
-    {
-      title: 'rule with good query (as KQL), with optional filter (as filter) and not total filter',
-      params: {
-        total: '',
-        good: 'http.response.status_code < 500',
-        filter: {
-          kqlQuery: '',
-          filters: [
-            {
-              $state: {
-                store: 'appState',
-              },
-              meta: {
-                field: 'host.name',
-                negate: false,
-                alias: null,
-                index: 'kbn-data-forge-fake_stack.admin-console-*-id',
-                disabled: false,
-                params: {
-                  query: 'admin-console.prod.001',
-                },
-                type: 'phrase',
-                key: 'host.name',
-              },
-              query: {
-                match_phrase: {
-                  'host.name': 'admin-console.prod.001',
-                },
-              },
-            },
-          ],
-        },
-      },
-    },
-    {
-      title:
-        'rule with good query (as KQL), with optional filter (as KQL) and total filter (as KQL)',
-      params: {
-        good: 'http.response.status_code < 500',
-        filter: 'host.name: admin-console.prod.001',
-        total: 'http.response.status_code: *',
-      },
-    },
-    {
-      title:
-        'rule with good query (as filter), with optional filter (as filter) and total filter (as KQL)',
-      params: {
-        good: {
-          kqlQuery: '',
-          filters: [
-            {
-              $state: {
-                store: 'appState',
-              },
-              meta: {
-                field: 'http.response.status_code',
-                negate: false,
-                alias: null,
-                index: 'kbn-data-forge-fake_stack.admin-console-*-id',
-                disabled: false,
-                params: {
-                  lt: '500',
-                },
-                type: 'range',
-                key: 'http.response.status_code',
-              },
-              query: {
-                range: {
-                  'http.response.status_code': {
-                    lt: '500',
-                  },
-                },
-              },
-            },
-          ],
-        },
-        filter: {
-          kqlQuery: '',
-          filters: [
-            {
-              $state: {
-                store: 'appState',
-              },
-              meta: {
-                field: 'host.name',
-                negate: false,
-                alias: null,
-                index: 'kbn-data-forge-fake_stack.admin-console-*-id',
-                disabled: false,
-                params: {
-                  query: 'admin-console.prod.001',
-                },
-                type: 'phrase',
-                key: 'host.name',
-              },
-              query: {
-                match_phrase: {
-                  'host.name': 'admin-console.prod.001',
-                },
-              },
-            },
-          ],
-        },
-        total: 'http.response.status_code: *',
-      },
-    },
-    {
-      title:
-        'rule with good query (as KQL), with optional filter (as KQL), total filter (as KQL) and group by two fields',
-      params: {
-        good: 'http.response.status_code < 500',
-        filter: 'host.name: admin-console.prod.001',
-        total: 'http.response.status_code: *',
-      },
-      groupBy: ['not_nested_1', 'not_nested_2'],
-      groupings: {
-        not_nested_1: 'authentication',
-        not_nested_2: 'blast-mail.co',
-      },
-    },
-  ];
-
-  test.each(testData)(
-    'should generate correct es query for $title',
-    ({ params, groupBy, groupings }) => {
-      expect(
-        getESQueryForLogRateAnalysis(params as KQLCustomIndicator['params'], groupBy, groupings)
-      ).toMatchSnapshot();
-    }
-  );
-});
