+++ x-pack/plugins/observability_solution/slo/public/components/slo/error_rate_chart/use_lens_definition.ts
@@ -13,7 +13,6 @@ import { ALL_VALUE, SLOWithSummaryResponse } from '@kbn/slo-schema';
 import moment from 'moment';
 import { v4 as uuidv4 } from 'uuid';
 import { SLO_DESTINATION_INDEX_PATTERN } from '../../../../common/constants';
-import { SloTabId } from '../../../pages/slo_details/components/slo_details';
 import { getLensDefinitionInterval } from './utils';
 
 export interface TimeRange {
@@ -26,24 +25,26 @@ export interface AlertAnnotation {
   total: number;
 }
 
+interface Props {
+  slo: SLOWithSummaryResponse;
+  threshold?: number;
+  dataTimeRange: TimeRange;
+  alertTimeRange?: TimeRange;
+  annotations?: AlertAnnotation[];
+  variant: 'success' | 'danger';
+}
+
 export function useLensDefinition({
   slo,
   threshold,
   dataTimeRange,
   alertTimeRange,
   annotations,
-  showErrorRateAsLine,
-  selectedTabId,
-}: {
-  slo: SLOWithSummaryResponse;
-  threshold: number;
-  dataTimeRange: TimeRange;
-  alertTimeRange?: TimeRange;
-  annotations?: AlertAnnotation[];
-  showErrorRateAsLine?: boolean;
-  selectedTabId?: SloTabId;
-}): TypedLensByValueInput['attributes'] {
+  variant,
+}: Props): TypedLensByValueInput['attributes'] {
   const { euiTheme } = useEuiTheme();
+
+  const lineColor = variant === 'danger' ? euiTheme.colors.danger : euiTheme.colors.success;
 
   const interval = getLensDefinitionInterval(dataTimeRange, slo);
 
@@ -88,18 +89,18 @@ export function useLensDefinition({
             layerId: '8730e8af-7dac-430e-9cef-3b9989ff0866',
             accessors: ['9f69a7b0-34b9-4b76-9ff7-26dc1a06ec14'],
             position: 'top',
-            seriesType: !!showErrorRateAsLine ? 'line' : 'area',
+            seriesType: 'line',
             showGridlines: false,
             layerType: 'data',
             xAccessor: '627ded04-eae0-4437-83a1-bbb6138d2c3b',
             yConfig: [
               {
                 forAccessor: '9f69a7b0-34b9-4b76-9ff7-26dc1a06ec14',
-                color: !!showErrorRateAsLine ? euiTheme.colors.primary : euiTheme.colors.danger,
+                color: lineColor,
               },
             ],
           },
-          ...(selectedTabId !== 'history'
+          ...(threshold !== undefined
             ? [
                 {
                   layerId: '34298f84-681e-4fa3-8107-d6facb32ed92',
@@ -398,66 +399,69 @@ export function useLensDefinition({
               incompleteColumns: {},
               sampling: 1,
             },
-            '34298f84-681e-4fa3-8107-d6facb32ed92': {
-              linkToLayers: [],
-              columns: {
-                '0a42b72b-cd5a-4d59-81ec-847d97c268e6X0': {
-                  label: `Part of ${threshold}x`,
-                  dataType: 'number',
-                  operationType: 'math',
-                  isBucketed: false,
-                  scale: 'ratio',
-                  params: {
-                    // @ts-ignore
-                    tinymathAst: {
-                      type: 'function',
-                      name: 'multiply',
-                      args: [
-                        {
-                          type: 'function',
-                          name: 'subtract',
-                          args: [1, slo.objective.target],
-                          location: {
-                            min: 1,
-                            max: 9,
+
+            ...(threshold !== undefined && {
+              '34298f84-681e-4fa3-8107-d6facb32ed92': {
+                linkToLayers: [],
+                columns: {
+                  '0a42b72b-cd5a-4d59-81ec-847d97c268e6X0': {
+                    label: `Part of ${threshold}x`,
+                    dataType: 'number',
+                    operationType: 'math',
+                    isBucketed: false,
+                    scale: 'ratio',
+                    params: {
+                      // @ts-ignore
+                      tinymathAst: {
+                        type: 'function',
+                        name: 'multiply',
+                        args: [
+                          {
+                            type: 'function',
+                            name: 'subtract',
+                            args: [1, slo.objective.target],
+                            location: {
+                              min: 1,
+                              max: 9,
+                            },
+                            text: `1 - ${slo.objective.target}`,
                           },
-                          text: `1 - ${slo.objective.target}`,
+                          threshold,
+                        ],
+                        location: {
+                          min: 0,
+                          max: 17,
                         },
-                        threshold,
-                      ],
-                      location: {
-                        min: 0,
-                        max: 17,
+                        text: `(1 - ${slo.objective.target}) * ${threshold}`,
                       },
-                      text: `(1 - ${slo.objective.target}) * ${threshold}`,
                     },
+                    references: [],
+                    customLabel: true,
                   },
-                  references: [],
-                  customLabel: true,
-                },
-                '0a42b72b-cd5a-4d59-81ec-847d97c268e6': {
-                  label: `${numeral(threshold).format('0.[00]')}x`,
-                  dataType: 'number',
-                  operationType: 'formula',
-                  isBucketed: false,
-                  scale: 'ratio',
-                  params: {
-                    // @ts-ignore
-                    formula: `(1 - ${slo.objective.target}) * ${threshold}`,
-                    isFormulaBroken: false,
+                  '0a42b72b-cd5a-4d59-81ec-847d97c268e6': {
+                    label: `${numeral(threshold).format('0.[00]')}x`,
+                    dataType: 'number',
+                    operationType: 'formula',
+                    isBucketed: false,
+                    scale: 'ratio',
+                    params: {
+                      // @ts-ignore
+                      formula: `(1 - ${slo.objective.target}) * ${threshold}`,
+                      isFormulaBroken: false,
+                    },
+                    references: ['0a42b72b-cd5a-4d59-81ec-847d97c268e6X0'],
+                    customLabel: true,
                   },
-                  references: ['0a42b72b-cd5a-4d59-81ec-847d97c268e6X0'],
-                  customLabel: true,
                 },
+                columnOrder: [
+                  '0a42b72b-cd5a-4d59-81ec-847d97c268e6',
+                  '0a42b72b-cd5a-4d59-81ec-847d97c268e6X0',
+                ],
+                sampling: 1,
+                ignoreGlobalFilters: false,
+                incompleteColumns: {},
               },
-              columnOrder: [
-                '0a42b72b-cd5a-4d59-81ec-847d97c268e6',
-                '0a42b72b-cd5a-4d59-81ec-847d97c268e6X0',
-              ],
-              sampling: 1,
-              ignoreGlobalFilters: false,
-              incompleteColumns: {},
-            },
+            }),
           },
         },
         indexpattern: {
