+++ x-pack/plugins/observability_solution/slo/public/pages/slo_details/components/history/slo_details_history.tsx
@@ -7,17 +7,18 @@
 import {
   EuiFlexGroup,
   EuiFlexItem,
-  EuiSpacer,
+  EuiPanel,
   EuiSuperDatePicker,
+  EuiTitle,
   OnRefreshProps,
   OnTimeChangeProps,
 } from '@elastic/eui';
 import DateMath from '@kbn/datemath';
+import { i18n } from '@kbn/i18n';
 import { SLOWithSummaryResponse } from '@kbn/slo-schema';
 import React, { useMemo, useState } from 'react';
-import { BurnRates } from '../../../../components/slo/burn_rate/burn_rates';
+import { ErrorRateChart } from '../../../../components/slo/error_rate_chart';
 import { useKibana } from '../../../../hooks/use_kibana';
-import { useBurnRateOptions } from '../../hooks/use_burn_rate_options';
 import { TimeBounds } from '../../types';
 import { EventsChartPanel } from '../events_chart_panel';
 import { HistoricalDataCharts } from '../historical_data_charts';
@@ -31,7 +32,6 @@ export interface Props {
 
 export function SLODetailsHistory({ slo, isAutoRefreshing, selectedTabId }: Props) {
   const { uiSettings } = useKibana().services;
-  const { burnRateOptions } = useBurnRateOptions(slo);
   const [start, setStart] = useState(`now-${slo.timeWindow.duration}`);
   const [end, setEnd] = useState('now');
 
@@ -55,7 +55,7 @@ export function SLODetailsHistory({ slo, isAutoRefreshing, selectedTabId }: Prop
   };
 
   return (
-    <>
+    <EuiFlexGroup direction="column" gutterSize="l">
       <EuiFlexGroup justifyContent="flexEnd">
         <EuiFlexItem
           grow
@@ -83,35 +83,40 @@ export function SLODetailsHistory({ slo, isAutoRefreshing, selectedTabId }: Prop
         </EuiFlexItem>
       </EuiFlexGroup>
 
-      <EuiSpacer size="l" />
+      <EuiPanel paddingSize="m" color="transparent" hasBorder data-test-subj="errorRatePanel">
+        <EuiFlexGroup direction="column" gutterSize="m">
+          <EuiFlexItem grow={false}>
+            <EuiTitle size="xs">
+              <h2>
+                {i18n.translate('xpack.slo.sloDetailsHistory.h2.errorRatePanelTitle', {
+                  defaultMessage: 'Error rate',
+                })}
+              </h2>
+            </EuiTitle>
+          </EuiFlexItem>
+          <ErrorRateChart
+            slo={slo}
+            dataTimeRange={range}
+            onBrushed={onBrushed}
+            variant={['VIOLATED', 'DEGRADING'].includes(slo.summary.status) ? 'danger' : 'success'}
+          />
+        </EuiFlexGroup>
+      </EuiPanel>
 
-      <EuiFlexGroup direction="column" gutterSize="l">
-        <EuiFlexItem>
-          <BurnRates
-            slo={slo}
-            isAutoRefreshing={isAutoRefreshing}
-            burnRateOptions={burnRateOptions}
-            selectedTabId={selectedTabId}
-            range={range}
-            onBrushed={onBrushed}
-          />
-        </EuiFlexItem>
-        <HistoricalDataCharts
-          slo={slo}
-          selectedTabId={selectedTabId}
-          isAutoRefreshing={isAutoRefreshing}
-          range={range}
-          onBrushed={onBrushed}
-        />
-        <EuiFlexItem>
-          <EventsChartPanel
-            slo={slo}
-            range={range}
-            selectedTabId={selectedTabId}
-            onBrushed={onBrushed}
-          />
-        </EuiFlexItem>
-      </EuiFlexGroup>
-    </>
+      <HistoricalDataCharts
+        slo={slo}
+        selectedTabId={selectedTabId}
+        isAutoRefreshing={isAutoRefreshing}
+        range={range}
+        onBrushed={onBrushed}
+      />
+
+      <EventsChartPanel
+        slo={slo}
+        range={range}
+        selectedTabId={selectedTabId}
+        onBrushed={onBrushed}
+      />
+    </EuiFlexGroup>
   );
 }
