+++ x-pack/plugins/observability_solution/slo/public/rules/register_burn_rate_rule_type.ts
@@ -73,7 +73,7 @@ export const registerBurnRateRuleType = (
     defaultActionMessage: sloBurnRateDefaultActionMessage,
     defaultRecoveryMessage: sloBurnRateDefaultRecoveryMessage,
     alertDetailsAppSection: lazyWithContextProviders(
-      lazy(() => import('../components/slo/burn_rate/alert_details/alert_details_app_section'))
+      lazy(() => import('../components/alert_details/alert_details_app_section'))
     ),
     priority: 100,
   });
